+++ jest.coverage.js
@@ -3,6 +3,14 @@ module.exports = {
   preset: 'jest-puppeteer',
   setupFilesAfterEnv: ['<rootDir>/test/setupTests.js'],
   collectCoverage: true,
-  collectCoverageFrom: ['<rootDir>/lib/Terminal.jsx'],
+  collectCoverageFrom: ['<rootDir>/src/**/*'],
+  coveragePathIgnorePatterns: [
+    '/node_modules/',
+    '/src/defs', // Defs cannot be tested
+    // These are covered by interactivity tests or other tests
+    'cleanArray.js',
+    'scrollHistory.js',
+    'sendCursorToEnd.js'
+  ],
   coverageDirectory: './coverage/'
 }