+++ src/Terminal.jsx
@@ -8,10 +8,14 @@ import TerminalMessage from './TerminalMessage'
 // Handlers
 import validateCommands from './handlers/validateCommands'
 import scrollHistory from './handlers/scrollHistory'
+import parseEOL from './handlers/parseEOL'
 
 // Definitions
 import sourceStyles from './defs/styles/Terminal'
 import types from './defs/types/Terminal'
+
+// Utils
+import commandExists from './utils/commandExists'
 
 export default class Terminal extends Component {
   constructor (props) {
@@ -22,52 +26,38 @@ export default class Terminal extends Component {
       history: [],
       historyPosition: null,
       previousHistoryPosition: null,
+      // TODO: Add prop-controlled enable/disable on the input
       processing: false
     }
 
     this.terminalRoot = React.createRef()
     this.terminalInput = React.createRef()
-
-    this.focusTerminal = this.focusTerminal.bind(this)
-    this.validateCommands = this.validateCommands.bind(this)
-    this.showWelcomeMessage = this.showWelcomeMessage.bind(this)
-    this.showHelp = this.showHelp.bind(this)
-    this.pushToStdout = this.pushToStdout.bind(this)
-    this.getStdout = this.getStdout.bind(this)
-    this.clearStdout = this.clearStdout.bind(this)
-    this.processCommand = this.processCommand.bind(this)
-    this.handleInput = this.handleInput.bind(this)
   }
 
   static propTypes = types
 
   /* istanbul ignore next: Covered by interactivity tests */
-  focusTerminal () {
+  focusTerminal = () => {
     // Only focus the terminal if text isn't being copied
     const isTextSelected = window.getSelection().type === 'Range'
     if (!isTextSelected) this.terminalInput.current.focus()
   }
 
   /* istanbul ignore next: Covered by interactivity tests */
-  scrollToBottom () {
+  scrollToBottom = () => {
     const rootNode = this.terminalRoot.current
 
     // This may look ridiculous, but it is necessary to decouple execution for just a millisecond in order to scroll all the way
     setTimeout(() => { rootNode.scrollTop = rootNode.scrollHeight }, 1)
   }
 
-  validateCommands () {
-    const validCommands = validateCommands(
-      this.props.commands,
-      this.showHelp,
-      this.clearStdout,
-      this.props.noDefaults
-    )
-
+  validateCommands = () => {
+    const { commands, noDefaults, ignoreCommandCase } = this.props
+    const validCommands = validateCommands(commands, this.showHelp, this.clearStdout, { noDefaults, ignoreCommandCase })
     this.setState({ commands: validCommands })
   }
 
-  showWelcomeMessage () {
+  showWelcomeMessage = () => {
     const msg = this.props.welcomeMessage
 
     if (typeof msg === 'boolean') this.pushToStdout('Welcome to the React terminal! Type \'help\' to get a list of commands.')
@@ -76,7 +66,7 @@ export default class Terminal extends Component {
   }
 
   /* istanbul ignore next: Covered by interactivity tests */
-  showHelp () {
+  showHelp = () => {
     const { commands } = this.state
 
     for (const c in commands) {
@@ -87,65 +77,98 @@ export default class Terminal extends Component {
     }
   }
 
-  pushToStdout (message, rawInput) {
-    const { stdout, history } = this.state
+  /**
+   * @param {String} message
+   * @param {Object} options {
+   *  rawInput: Raw input from the terminal (For history),
+   *  isEcho: For distinguishing echo messages (Exemption from message styling)
+   * }
+   */
+  pushToStdout = (message, options) => {
+    const { stdout } = this.state
+    stdout.push({ message, isEcho: options?.isEcho || false })
 
-    stdout.push(message)
-
-    /* istanbul ignore if: Covered by interactivity tests */
-    if (rawInput) { // Only supplied if history is enabled
-      history.push(rawInput)
-      this.setState({ stdout: stdout, history: history, historyPosition: null })
-    } else {
-      this.setState({ stdout: stdout })
-    }
+    /* istanbul ignore next: Covered by interactivity tests */
+    if (options?.rawInput) this.pushToHistory(options.rawInput)
+    this.setState({ stdout: stdout })
   }
 
-  getStdout () {
-    return this.state.stdout.map((line, i) => <TerminalMessage key={i} content={line} />)
+  /**
+   * @param {String} rawInput Raw command input from the terminal
+   */
+  /* istanbul ignore next: Covered by interactivity tests */
+  pushToHistory = rawInput => {
+    const { history } = this.state
+    history.push(rawInput)
+    this.setState({ history: history, historyPosition: null })
+  }
+
+  getStdout = () => {
+    // Parse EOL if it isn't disabled
+    const stdout = !this.props.noNewlineParsing ? parseEOL(this.state.stdout) : this.state.stdout
+
+    return stdout.map((line, i) => {
+      return <TerminalMessage
+        key={i}
+        content={line.message}
+        dangerMode={this.props.dangerMode}
+        className={!line.isEcho ? this.props.messageClassName : /* istanbul ignore next: Covered by interactivity tests */ undefined}
+        style={!line.isEcho ? this.props.messageStyle : /* istanbul ignore next: Covered by interactivity tests */ undefined}
+      />
+    })
   }
 
   /* istanbul ignore next: Covered by interactivity tests */
-  clearStdout () {
+  clearStdout = () => {
     this.setState({ stdout: [] })
   }
 
   /* istanbul ignore next: Covered by interactivity tests */
-  clearInput () {
+  clearInput = () => {
     this.setState({ historyPosition: null })
     this.terminalInput.current.value = ''
   }
 
   /* istanbul ignore next: Covered by interactivity tests */
-  processCommand () {
+  processCommand = () => {
     this.setState({ processing: true }, () => {
       // Initialise command result object
       const commandResult = { command: null, args: [], rawInput: null, result: null }
       const rawInput = this.terminalInput.current.value
 
-      if (!this.props.noAutomaticStdout) {
-        if (!this.props.noHistory) this.pushToStdout(`${this.props.promptLabel || '$'} ${rawInput}`, rawInput)
-        else this.pushToStdout(`${this.props.promptLabel || '$'} ${rawInput}`)
+      if (!this.props.noHistory) this.pushToHistory(rawInput)
+
+      if (!this.props.noEchoBack) {
+        // Mimic native terminal by echoing command back
+        // Also exempt it from message since it should not really be a message despite behaving like one
+        // Containing it in a span to allow JSX values in the prompt label
+        const echo = <span>{this.props.promptLabel || '$'} {rawInput}</span>
+        this.pushToStdout(echo, { isEcho: true })
       }
 
       if (rawInput) {
         const input = rawInput.split(' ')
-        const command = input.splice(0, 1)[0] // Removed portion is returned...
+        const rawCommand = input.splice(0, 1)[0] // Removed portion is returned...
         const args = input // ...and the rest can be used
 
         commandResult.rawInput = rawInput
-        commandResult.command = command
+        commandResult.command = rawCommand
         commandResult.args = args
 
-        const cmdObj = this.state.commands[command]
+        const { exists, command } = commandExists(this.state.commands, rawCommand, this.props.ignoreCommandCase)
 
-        if (!cmdObj) this.pushToStdout(this.props.errorText ? this.props.errorText.replace(/\[command\]/gi, command) : `Command '${command}' not found!`)
-        else {
-          const res = cmdObj.fn(...args)
+        if (!exists) {
+          this.pushToStdout(this.props.errorText
+            ? this.props.errorText.replace(/\[command\]/gi, command)
+            : `Command '${rawCommand}' not found!`
+          )
+        } else {
+          const cmd = this.state.commands[command]
+          const res = cmd.fn(...args)
 
           this.pushToStdout(res)
           commandResult.result = res
-          if (cmdObj.explicitExec) cmdObj.fn(...args)
+          if (cmd.explicitExec) cmd.fn(...args)
         }
       }
 
@@ -158,21 +181,22 @@ export default class Terminal extends Component {
   }
 
   /* istanbul ignore next: Covered by interactivity tests */
-  scrollHistory (direction) {
-    const toUpdate = scrollHistory(
-      direction,
-      this.state.history,
-      this.state.historyPosition,
-      this.state.previousHistoryPosition,
-      this.terminalInput,
-      this.props.noAutomaticStdout
-    )
+  scrollHistory = direction => {
+    const { history, historyPosition, previousHistoryPosition } = this.state
 
-    this.setState(toUpdate)
+    const toUpdate = scrollHistory(direction, {
+      history,
+      historyPosition,
+      previousHistoryPosition,
+      terminalInput: this.terminalInput
+    })
+
+    // Only update if there is something to update
+    if (toUpdate) this.setState(toUpdate)
   }
 
   /* istanbul ignore next: Covered by interactivity tests */
-  handleInput (event) {
+  handleInput = event => {
     switch (event.key) {
       case 'Enter': this.processCommand(); break
       case 'ArrowUp': this.scrollHistory('up'); break
@@ -204,14 +228,14 @@ export default class Terminal extends Component {
     return (
       <div
         ref={this.terminalRoot}
-        name={'react-console-emulator'}
+        name='react-console-emulator'
         className={this.props.className}
         style={styles.container}
         onClick={this.focusTerminal}
       >
         {/* Content */}
         <div
-          name={'react-console-emulator__content'}
+          name='react-console-emulator__content'
           className={this.props.contentClassName}
           style={styles.content}
         >
@@ -219,13 +243,13 @@ export default class Terminal extends Component {
           {this.getStdout()}
           {/* Input area */}
           <div
-            name={'react-console-emulator__inputArea'}
+            name='react-console-emulator__inputArea'
             className={this.props.inputAreaClassName}
             style={styles.inputArea}
           >
             {/* Prompt label */}
             <span
-              name={'react-console-emulator__promptLabel'}
+              name='react-console-emulator__promptLabel'
               className={this.props.promptLabelClassName}
               style={styles.promptLabel}
             >
@@ -234,15 +258,15 @@ export default class Terminal extends Component {
             {/* Input */}
             <input
               ref={this.terminalInput}
-              name={'react-console-emulator__input'}
+              name='react-console-emulator__input'
               className={this.props.inputClassName}
               style={styles.input}
               onKeyDown={this.handleInput}
-              type={'text'}
-              autoComplete={'off'}
+              type='text'
+              autoComplete='off'
               disabled={
-                this.props.disableOnProcess &&
-                /* istanbul ignore next: Covered by interactivity tests */ this.state.processing
+                this.props.disabled ||
+                (this.props.disableOnProcess && /* istanbul ignore next: Covered by interactivity tests */ this.state.processing)
               }
             />
           </div>