+++ src/defs/types/TerminalMessage.js
@@ -1,0 +1,8 @@
+import PropTypes from 'prop-types'
+
+export default {
+  content: PropTypes.node,
+  style: PropTypes.object,
+  className: PropTypes.string,
+  dangerMode: PropTypes.bool
+}