+++ test/Terminal.test.js
@@ -4,18 +4,7 @@ import React from 'react'
 import { shallow, mount, render } from 'enzyme'
 import skipIf from 'skip-if'
 
-import * as src from '../lib/Terminal'
-import * as prod from '../dist/Terminal'
-
-let Terminal
-
-if (process.env.PROD_RUN) {
-  console.log('Performing production run, testing version from "dist".')
-  Terminal = prod.default
-} else {
-  console.log('Performing development run, testing version from "lib".')
-  Terminal = src.default
-}
+import Terminal from '../src/Terminal'
 
 const skipIfCI = skipIf(process.env.CI)
 
@@ -57,6 +46,22 @@ const newDefaultCommands = {
   }
 }
 
+const evilNameCommands = {
+  'te&st!': {
+    fn: () => ''
+  }
+}
+
+const duplicateCommands = {
+  help: {
+    fn: () => 'Help!'
+  }
+}
+
+const invalidCommands = {
+  noFn: {}
+}
+
 beforeAll(async () => {
   await page.goto('http://localhost:8000')
 })
@@ -75,7 +80,7 @@ describe('Terminal HTML structure', () => {
 
 describe('Terminal welcome messages', () => {
   it('Displays default welcome', () => {
-    const wrapper = mount(<Terminal commands={commands} welcomeMessage={true}/>)
+    const wrapper = mount(<Terminal commands={commands} welcomeMessage/>)
     const content = wrapper.find('[name="react-console-emulator__content"]')
 
     expect(content.childAt(0).text()).toBe('Welcome to the React terminal! Type \'help\' to get a list of commands.')
@@ -84,7 +89,7 @@ describe('Terminal welcome messages', () => {
   })
 
   it('Displays custom welcomes', () => {
-    const wrapperSingle = mount(<Terminal commands={commands} welcomeMessage={'test'}/>)
+    const wrapperSingle = mount(<Terminal commands={commands} welcomeMessage='test'/>)
     const wrapperMulti = mount(<Terminal commands={commands} welcomeMessage={['test', 'test2']}/>)
 
     const singleContent = wrapperSingle.find('[name="react-console-emulator__content"]')
@@ -99,6 +104,16 @@ describe('Terminal welcome messages', () => {
 
     wrapperSingle.unmount()
     wrapperMulti.unmount()
+  })
+
+  it('Renders HTML in welcome messages', () => {
+    const wrapper = mount(<Terminal commands={commands} welcomeMessage='<span color="red">test</span>' dangerMode/>)
+
+    const content = wrapper.find('[name="react-console-emulator__content"]')
+
+    expect(content.childAt(0).html()).toBe('<div style="line-height: 21px;"><span color="red">test</span></div>')
+
+    wrapper.unmount()
   })
 })
 
@@ -115,12 +130,47 @@ describe('Terminal functionality', () => {
   })
 
   it('Registers new default commands', () => {
-    const wrapper = mount(<Terminal commands={newDefaultCommands} noDefaults={true}/>)
+    const wrapper = mount(<Terminal commands={newDefaultCommands} noDefaults/>)
 
     expect(wrapper.state().commands).toHaveProperty('validation')
     expect(wrapper.state().commands).toHaveProperty('help')
 
     wrapper.unmount()
+  })
+
+  it('Parses newlines (But not when disabled)', () => {
+    const wrapperEnabled = mount(<Terminal commands={commands} welcomeMessage={'split1\\nsplit2'}/>)
+    const wrapperDisabled = mount(<Terminal commands={commands} welcomeMessage={'split1\\nsplit2'} noNewlineParsing/>)
+
+    const enabledContent = wrapperEnabled.find('[name="react-console-emulator__content"]')
+    const disabledContent = wrapperDisabled.find('[name="react-console-emulator__content"]')
+
+    // Splits with parsing enabled...
+    expect(enabledContent.childAt(0).text()).toBe('split1')
+    expect(enabledContent.childAt(1).text()).toBe('split2')
+
+    // ...and doesn't with parsing disabled
+    expect(disabledContent.childAt(0).text()).toBe('split1\\nsplit2')
+
+    wrapperEnabled.unmount()
+    wrapperDisabled.unmount()
+  })
+})
+
+describe('Terminal command validator', () => {
+  it('Throws an error when evil names are submitted to the regex', () => {
+    expect(() => shallow(<Terminal commands={evilNameCommands} ignoreCommandCase/>))
+      .toThrowError(/Command name '\S+' is invalid; command names can only contain latin characters \(A-Z\), numbers \(0-9\) and dashes\/underscores \(- or _\)/)
+  })
+
+  it('Throws an error when default commands are being overriden without the noDefaults property being set', () => {
+    expect(() => shallow(<Terminal commands={duplicateCommands}/>))
+      .toThrowError(/Attempting to override existing command '\S+'; please only supply one definition of a certain command, or set the noDefaults property to enable overriding of existing commands/)
+  })
+
+  it('Throws an error when commands with an invalid shape are submitted', () => {
+    expect(() => shallow(<Terminal commands={invalidCommands}/>))
+      .toThrowError(/'fn' property of command '\S+' is invalid; expected 'function', got '\S+'/)
   })
 })
 
@@ -187,7 +237,6 @@ describe('Terminal user interactivity', () => {
     await page.keyboard.press('ArrowUp')
     expect(await getInputValue()).toBe('echo test')
     await page.keyboard.press('ArrowDown')
-    await page.keyboard.press('ArrowDown') // Temp workaround
     expect(await getInputValue()).toBe('')
   })
 })