+++ CHANGELOG.md
@@ -1,18 +1,59 @@
+# 4.0.0
+
+This update brings with it the biggest batch of changes to react-console-emulator since its initial release. Changes include a move to a new location, major restructurations of the repository and the internal logic, a whole bunch of changes, improvements, new features and of course some bug fixes.
+
+### Breaking changes
+
+Moved repositories; js-rcon/react-console-emulator is now linuswillner/react-console-emulator!
+
+Renamed `noAutomaticStdout` prop to `noEchoBack` for added clarity.
+
+### Main changes
+
+Terminal message styling is here! You can now re-style the messages output by the terminal (Including echoes, optionally with the `styleEchoBack` prop) using the `messageStyle` and `messageClassName` props ([#518](https://github.com/linuswillner/react-console-emulator/issues/518)).
+
+JSX prompt labels! Prompt labels now support elements instead of just plain old strings ([#210](https://github.com/linuswillner/react-console-emulator/issues/210)).
+
+Newline parsing is now possible! The terminal can now parse newline characters in terminal messages - anything with a \n character in it will be rendered as a separate line in the response message. This does of course not apply to command back-echoes. This behaviour can also be disabled, if desired, using the `noNewlineParsing` prop ([#519](https://github.com/linuswillner/react-console-emulator/issues/519)).
+
+
+Case-insensitive command matching! You can now supply the `ignoreCommandCase` prop to allow matching commands even when their casing is not correct. Do note that for security reasons, enabling case-insensitive command matching restricts command names to letters, numbers and dashes/underscores ([#415](https://github.com/linuswillner/react-console-emulator/issues/415)).
+
+### Other changes
+
+The acceptance of terminal input can now be controlled with a prop. Setting the `disabled` prop will enable/disable character input to the terminal.
+
+Fixed an extremely long-running bug where, when the history only had one unit in it, the terminal would not correctly reset the input value and would show the same value twice. This has now finally been fixed after having been in the library for over two years.
+
+Completely reworked demo page - the old one was really drab and plain, so I took the time to make it a lot more pretty and informative.
+
+A full run-down of the prop changes is as follows:
+```diff
+- noAutomaticStdout
++ noEchoBack
++ noNewlineParsing
++ ignoreCommandCase
++ styleEchoBack
++ messageStyle
++ messageClassName
++ disabled
+```
+
 # 3.0.4
 
-Fixed a bug preventing users from selecting text in the terminal output ([#414](https://github.com/js-rcon/react-console-emulator/issues/414)).
+Fixed a bug preventing users from selecting text in the terminal output ([#414](https://github.com/linuswillner/react-console-emulator/issues/414)).
 
 # 3.0.3
 
-Removed redundant `stringify-object` dependency to properly enable command re-validation based on raw objects alone. This was supposed to have been fixed in 3.0.2, but due to a mishap the old validation was left dangling. This has now been fixed ([#411](https://github.com/js-rcon/react-console-emulator/issues/411)).
+Removed redundant `stringify-object` dependency to properly enable command re-validation based on raw objects alone. This was supposed to have been fixed in 3.0.2, but due to a mishap the old validation was left dangling. This has now been fixed ([#411](https://github.com/linuswillner/react-console-emulator/issues/411)).
 
 # 3.0.2
 
-Fixed command re-validation reliability issues relating to source-identical commands ([#35](https://github.com/js-rcon/react-console-emulator/issues/35)).
+Fixed command re-validation reliability issues relating to source-identical commands ([#35](https://github.com/linuswillner/react-console-emulator/issues/35)).
 
 # 3.0.1
 
-Fixed input outline showing on Safari ([#258](https://github.com/js-rcon/react-console-emulator/pull/258)) ([Herve07h22](https://github.com/Herve07h22)).
+Fixed input outline showing on Safari ([#258](https://github.com/linuswillner/react-console-emulator/pull/258)) ([Herve07h22](https://github.com/Herve07h22)).
 
 # 3.0.0
 
@@ -36,7 +77,7 @@ The following style-related props have been removed and replaced:
 + inputStyle
 ```
 
-Instead of having a unique prop for styling only some parts of the terminal elements, every style aspect is now overridable. Any styles defines in [lib/defs/styles/Terminal.js](lib/defs/styles/Terminal.js) can be overridden via the new props. See [the guide](docs/CONFIG.md#re-styling) for more details.
+Instead of having a unique prop for styling only some parts of the terminal elements, every style aspect is now overridable. Any styles defines in [src/defs/styles/Terminal.js](src/defs/styles/Terminal.js) can be overridden via the new props. See [the guide](docs/CONFIG.md#re-styling) for more details.
 
 ### Other changes
 
@@ -72,19 +113,19 @@ Enabled module transpilation to widen the support amongst Node versions for dist
 
 # 1.7.2
 
-Re-added Babel into build flow in a different format to improve compatibility ([#39, comment](https://github.com/js-rcon/react-console-emulator/issues/39#issuecomment-440973765)).
+Re-added Babel into build flow in a different format to improve compatibility ([#39, comment](https://github.com/linuswillner/react-console-emulator/issues/39#issuecomment-440973765)).
 
 # 1.7.1
 
-Removed Babel from the build flow in order to allow the inclusion of the helper files ([#39](https://github.com/js-rcon/react-console-emulator/issues/39)).
+Removed Babel from the build flow in order to allow the inclusion of the helper files ([#39](https://github.com/linuswillner/react-console-emulator/issues/39)).
 
 # 1.7.0
 
 Internal refactoring for better maintainability.
 
-Added default-enabled automatic scrolling to the bottom of the terminal when a command is run ([#36](https://github.com/js-rcon/react-console-emulator/issues/36)).
+Added default-enabled automatic scrolling to the bottom of the terminal when a command is run ([#36](https://github.com/linuswillner/react-console-emulator/issues/36)).
 
-Added command callback support to run a function each time a command is executed ([#36](https://github.com/js-rcon/react-console-emulator/issues/36)).
+Added command callback support to run a function each time a command is executed ([#36](https://github.com/linuswillner/react-console-emulator/issues/36)).
 
 Added `noAutoScroll` and `commandCallback` props.